Update 021-mandelbrot-p5.md
Made a mandelbrot-explorer that allows you to scroll in and out and investigate yourself. Number of iterations can be changes with a select menu. It generates faster, low-resolution images while exploring and will render a high resolution image when the user stops interacting. Includes a few notable locations that can be auto- zoomed to by clicking on the named buttons.